+++ backend/src/migrations/migrate-book_cart.js
@@ -21,7 +21,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       unitPrice: {
-        type: Sequelize.DOUBLE
+        type: Sequelize.DECIMAL(10, 2)
       },
       createdAt: {
         allowNull: false,
